+++ src/main.js
@@ -6,6 +6,7 @@ import Vue from 'vue'
 import BootstrapVue from 'bootstrap-vue'
 import App from './App'
 import router from './router/index'
+import axios from './backend/vue-axios'
 
 // todo
 // cssVars()
@@ -16,6 +17,7 @@ Vue.use(BootstrapVue)
 new Vue({
   el: '#app',
   router,
+  axios,
   template: '<App/>',
   components: {
     App