+++ changelog.md
@@ -7,3 +7,5 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Template Core-UI-pro.
 - changelog.md.
+- Login is added.
+- Demo module Books. Index of books. Public.